Use a enum class for QoS type
Convert to integer only in the interface with Paho MQTT C. This
ensures that QoS will be used consciously. And the user will not
provide an arbitrary integer instead of a QoS value.

      +-----------------+
      |                 |
      |  Paho MQTT C++  | <-- Use only QoS enumeration
      |                 |
      +-----------------+ <-- Conversion between QoS enum and int
      |                 |
      |   Paho MQTT C   | <-- Deal with interger QoS
      |                 |
      +-----------------+

The mqtt::QoS enumerators have sizeof(int) because an array of QoS
might be cast to an array of integers. Thus, each element must have
the same size.

Signed-off-by: Guilherme Maciel Ferreira <guilherme.maciel.ferreira@gmail.com>