Use the same semantic as Paho MQTT C for invalid strings
When connecting with a SSL enabled server, the Paho client issues the
following errors:

  SSL routines:SSL_CTX_use_certificate_chain_file:system lib:ssl_rsa.c:701
  SSL routines:FILE_CTRL:system lib:bss_file.c:400

The error is caused because the SSLSocket_createContext() function
uses the MQTTClient_SSLOptions's members (keyStore, trustStore and
enabledCipherSuites) even if they are empty strings. The only way the
function doesn't use a particular member is setting it to NULL.

However, the ssl_options's constructors and assignment operators
initialize the members keyStore, trustStore and enabledCipherSuites
with empty strings (std::strings::c_str()).

So, despite setting only the trustStore, the user ends up passing
all members as valid, not NULL, strings. Then, SSLSocket_createContext()
calls SSL_CTX_use_certificate_chain_file() using an empty keyStore.

The call stack to the problem is:

  ----------- Paho MQTT C++ ------------
    async_client::connect()
  ----------- Paho MQTT C ------------
    MQTTAsync_connect()
    MQTTAsync_receiveThread()
    MQTTAsync_cycle()
    MQTTAsync_connecting()
    SSLSocket_setSocketForSSL()
    SSLSocket_createContext()
  ----------- OpenSSL -----------------
    SSL_CTX_use_certificate_chain_file()
    BIO_read_filename()

In order to solve the problem, the ssl_options class only inititalizes
a MQTTAsync_SSLOptions member if the string is not empty. So both C and
C++ libraries agree that NULL is a invalid string.

Signed-off-by: Guilherme Maciel Ferreira <guilherme.maciel.ferreira@gmail.com>