Fix invalid memory reference in iclient_persistence::persistence_put()
In the method iclient_persistence::persistence_put(), the "buf" string's
internal array (1) is referenced by "p" (2). However, "buf" is not in the
stack anymore when it is later referenced indirectly by the method
iclient_persistence::put() (3).

  {
    std::string buf; // (1)
    ...
    p = std::make_shared<persistence_wrapper>(&buf[0], buf.size()); // (2)
  }
  static_cast<iclient_persistence*>(handle)->put(key, p); // (3)

The simplest solution is to put "buf" in the same scope as the last method
that references it (3). So we don't risk to get "buf" removed from stack before
it is used by the last time (3):

  std::string buf;
  {
    ...
    p = std::make_shared<persistence_wrapper>(&buf[0], buf.size());
  }
  static_cast<iclient_persistence*>(handle)->put(key, p);

Signed-off-by: Guilherme Maciel Ferreira <guilherme.maciel.ferreira@gmail.com>