Fix invalid memory reference in iclient_persistence::persistence_put()
In the method iclient_persistence::persistence_put(), the "buf" string's
internal array (1) is referenced by "p" (2). However, "buf" is not in the
stack anymore when it is later referenced indirectly by the method
iclient_persistence::put() (3).

  {
    std::string buf; // (1)
    ...
    p = std::make_shared<persistence_wrapper>(&buf[0], buf.size()); // (2)
  }
  static_cast<iclient_persistence*>(handle)->put(key, p); // (3)

This solution replaces naked uint8_t pointers in persistence_wrapper by a
std::vector<uint8_t>. So the persistence_wrapper holds a copy of the data,
not a pointer to it. Thus, the data life spaw in memory is the tied to the
persistence_wrapper.

Signed-off-by: Guilherme Maciel Ferreira <guilherme.maciel.ferreira@gmail.com>