CMake builds against Paho C development tree
Allow CMake to work with Paho MQTT C development builds. The headers and the
library are not in standard directories, "include" and "lib", respectively.
Instead, in the Paho MQTT C development building, the headers are in
"${PAHO_MQTT_C_PATH}/src" and the libraries are in "${PAHO_MQTT_C_PATH}/build/output".

Tested with
- CMake version 2.8.12
- GNU GCC 4.8.2

Signed-off-by: Guilherme Maciel Ferreira <guilherme.maciel.ferreira@gmail.com>