Autotools: automatically detect if the SSL support is available
Remove the --with(out)-ssl argument. The Autotools no longer rely on the
user to decide whether or not to use SSL support. Now the decision depends
on whether there is a Paho MQTT C library with support for SSL.

Enable the SSL if there is a Paho MQTT C available with support for SSL. If
there is only a Paho MQTT C library without SSL support, disable the SSL.

The rationale behind the automatic decision is because the user already
decided about SSL support when he built the Paho MQTT C.

Change Paho MQTT C installation through Travis. It install the version without
SSL in the system directory. Otherwise, the SSL-enabled Paho MQTT C is always
used once the system directories have precedence over the --with-paho-mqtt-c
argument.

Signed-off-by: Guilherme Maciel Ferreira <guilherme.maciel.ferreira@gmail.com>