Fix nodiscard warnings in sync client
In later C++ versions, std::async is marked nodiscard. See https://cplusplus.github.io/LWG/issue2856.
The code is correct in that it is properly blocking for the result, but it will now warn since this is considered "not the way".
Adding .wait() to the end of the calls, which explicitly waits on the future, should be the correct fix.

Signed-off-by: Daniel Berlin <dberlin@dberlin.org>